+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SummerFiles/Auto_IMU.java
@@ -6,7 +6,7 @@
 //
 // Also uses IMU to drive in a straight line when not avoiding an obstacle.
 
-package org.firstinspires.ftc.teamcode.ConceptTests;
+package org.firstinspires.ftc.teamcode.SummerFiles;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
@@ -17,8 +17,6 @@ import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
-import org.firstinspires.ftc.robotcore.external.navigation.Position;
-import org.firstinspires.ftc.robotcore.external.navigation.Velocity;
 import org.firstinspires.ftc.teamcode.SourceFiles.Trobot;
 
 @Autonomous(name="Drive Avoid Imu", group="Exercises")
